# Reduce the main thread blocking in clients cron