The main thread needs to check clients in every cron iteration. During this check, the corresponding I/O threads must not operate on these clients to avoid data-race. As a result, the main thread is blocked until the I/O threads finish processing and are suspended, allowing the main thread to proceed with client checks.

Since the main thread's resources are more valuable than those of I/O threads in Redis, this blocking behavior should be avoided. To address this, the I/O threads check during their cron whether any of their maintained clients need to be inspected by the main thread. If so, the I/O threads send those clients to the main thread for processing, then the main thread runs cron jobs for these clients.

In addition, an always-active client might not be in thread->clients, so before processing the client’s command, we also check whether the client has skipped running its cron job for over 1 second. If it has, we run the cron job for the client.

The main thread does not need to actively pause the IO threads, thus avoiding potential blocking behavior, fixes https://github.com/redis/redis/issues/13885

Besides, this approach also can let all clients run cron task in a second, but before, we pause IO threads in multiple batches when there are more than 8 IO threads, that may cause some clients are not be processed in a second.

### Testing
The following are the test results for a large number of clients and a small number of clients:

./memtier_benchmark --data-size 521 --ratio 1:1 --key-pattern R:R --key-minimum=1 --key-maximum 3000000 --test-time 60 -c 500 -t 20 --hide-histogram -x 3

**After the modification.**
```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       210691.07          ---          ---        23.72842        19.58300        97.79100       194.55900    116997.51 
Gets       210608.24      1270.93    209337.31        23.73433        19.58300        98.30300       196.60700      8806.97 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     421299.31      1270.93    209337.31        23.73138        19.58300        97.79100       195.58300    125804.48 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       197954.78          ---          ---        25.63934        20.60700       114.68700       237.56700    109923.62 
Gets       197871.99      1181.70    196690.29        25.64540        20.60700       114.68700       238.59100      8267.79 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     395826.77      1181.70    196690.29        25.64237        20.60700       114.68700       238.59100    118191.41 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       203967.23          ---          ---        24.63106        19.58300       116.73500       288.76700    113263.11 
Gets       203884.16      1219.48    202664.68        24.63038        19.58300       116.73500       288.76700      8520.10 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     407851.39      1219.48    202664.68        24.63072        19.58300       116.73500       288.76700    121783.21
```

**Before the modification.**
```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       141862.52          ---          ---        35.24939        26.36700       221.18300       505.85500     78774.45 
Gets       141780.65       832.16    140948.50        35.25986        26.36700       221.18300       507.90300      5914.50 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     283643.17       832.16    140948.50        35.25463        26.36700       221.18300       505.85500     84688.95 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       119752.08          ---          ---        41.74563        34.30300       148.47900       251.90300     66496.93 
Gets       119670.15       831.96    118838.19        41.77178        34.30300       148.47900       251.90300      5058.01 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     239422.23       831.96    118838.19        41.75870        34.30300       148.47900       251.90300     71554.94 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       129590.86          ---          ---        38.57358        30.59100       167.93500       358.39900     71960.38 
Gets       129508.82       831.98    128676.83        38.59878        30.59100       167.93500       360.44700      5439.06 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     259099.67       831.98    128676.83        38.58618        30.59100       167.93500       358.39900     77399.44
```



./memtier_benchmark --data-size 521 --ratio 1:1 --key-pattern R:R --key-minimum=1 --key-maximum 3000000 --test-time 60 -c 50 -t 13 --hide-histogram -x 3

**After the modification.**

```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       298925.06          ---          ---         1.08710         0.88700         4.22300        10.36700    165994.53 
Gets       298919.71      1387.68    297532.03         1.08452         0.88700         4.22300        10.23900     12278.53 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     597844.77      1387.68    297532.03         1.08581         0.88700         4.22300        10.30300    178273.06 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       243856.22          ---          ---         1.33211         0.91100         7.77500        15.35900    135414.62 
Gets       243850.42      2410.64    241439.77         1.33055         0.91100         7.80700        15.42300     10671.36 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     487706.63      2410.64    241439.77         1.33133         0.91100         7.80700        15.42300    146085.98 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       271360.16          ---          ---         1.19724         0.89500         6.04700        13.05500    150687.73 
Gets       271354.62      2132.75    269221.87         1.19530         0.88700         6.01500        12.99100     11593.31 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     542714.78      2132.75    269221.87         1.19627         0.89500         6.04700        13.05500    162281.03
```

**Before the modification.**

```
BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       271636.14          ---          ---         1.19553         0.96700         4.70300        12.67100    150841.13 
Gets       271630.91      1160.77    270470.14         1.19406         0.95900         4.70300        12.73500     11106.64 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     543267.05      1160.77    270470.14         1.19479         0.95900         4.70300        12.67100    161947.76 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       198846.37          ---          ---         1.63286         1.09500         9.02300        24.19100    110420.10 
Gets       198841.06      1671.91    197169.15         1.63212         1.08700         9.02300        24.44700      8551.63 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     397687.43      1671.91    197169.15         1.63249         1.09500         9.02300        24.31900    118971.73 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       234881.84          ---          ---         1.38263         1.00700         6.94300        17.53500    130431.09 
Gets       234876.61      1637.88    233238.73         1.38114         1.00700         6.94300        17.66300      9928.69 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     469758.45      1637.88    233238.73         1.38188         1.00700         6.94300        17.66300    140359.78 
```