@egarevoc999 2025-03-31
> @wclmxxs i didn't reproduce your benchmark report from my local, almost nothing changed. What environment are you running the benchmark?
> 
> unstable
> 
> ```
> BEST RUN RESULTS
> ============================================================================================================================
> Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
> ----------------------------------------------------------------------------------------------------------------------------
> Sets       619071.91          ---          ---         0.52565         0.50300         1.00700         1.27900    343773.86 
> Gets       619066.74      5868.39    613198.35         0.52469         0.50300         1.00700         1.27100     26959.66 
> Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
> Totals    1238138.65      5868.39    613198.35         0.52517         0.50300         1.00700         1.27900    370733.52 
> 
> 
> WORST RUN RESULTS
> ============================================================================================================================
> Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
> ----------------------------------------------------------------------------------------------------------------------------
> Sets       607008.83          ---          ---         0.53590         0.51900         1.02300         1.27900    337075.18 
> Gets       607003.82     11060.60    595943.21         0.53530         0.51100         1.02300         1.27900     29149.95 
> Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
> Totals    1214012.65     11060.60    595943.21         0.53560         0.51100         1.02300         1.27900    366225.12 
> 
> 
> AGGREGATED AVERAGE RESULTS (3 runs)
> ============================================================================================================================
> Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
> ----------------------------------------------------------------------------------------------------------------------------
> Sets       612717.32          ---          ---         0.53101         0.51100         1.01500         1.28700    340245.11 
> Gets       612712.05      9375.74    603336.31         0.53022         0.51100         1.01500         1.28700     28508.59 
> Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
> Totals    1225429.37      9375.74    603336.31         0.53061         0.51100         1.01500         1.28700    368753.70 
> ```
> 
> this PR:
> 
> ```
> BEST RUN RESULTS
> ============================================================================================================================
> Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
> ----------------------------------------------------------------------------------------------------------------------------
> Sets       619601.55          ---          ---         0.52529         0.50300         0.99900         1.25500    344067.92 
> Gets       619596.05      5869.68    613726.37         0.52413         0.50300         0.99900         1.24700     26980.81 
> Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
> Totals    1239197.60      5869.68    613726.37         0.52471         0.50300         0.99900         1.24700    371048.73 
> 
> 
> WORST RUN RESULTS
> ============================================================================================================================
> Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
> ----------------------------------------------------------------------------------------------------------------------------
> Sets       612748.33          ---          ---         0.53091         0.51100         1.01500         1.24700    340262.33 
> Gets       612742.62     11272.50    601470.12         0.53026         0.51100         1.01500         1.24700     29480.36 
> Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
> Totals    1225490.95     11272.50    601470.12         0.53058         0.51100         1.01500         1.24700    369742.69 
> 
> 
> AGGREGATED AVERAGE RESULTS (3 runs)
> ============================================================================================================================
> Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
> ----------------------------------------------------------------------------------------------------------------------------
> Sets       615925.93          ---          ---         0.52824         0.51100         1.00700         1.24700    342026.85 
> Gets       615920.38      9494.24    606426.14         0.52745         0.51100         1.00700         1.24700     28693.34 
> Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
> Totals    1231846.31      9494.24    606426.14         0.52784         0.51100         1.00700         1.24700    370720.19 
> ```

Is it my environment? I don't seem to be much different now.
But you can try changing client from 50 to 500:
memtier_benchmark --data-size 521 --ratio 1:1 --key-pattern R:R --key-minimum=1 --key-maximum 3000000 --test-time 60 -c 500 -t 13 --hide-histogram -x 3 --port 6378