@wclmxxs 2025-04-21
@ShooterIT 
I was concerned that environmental fluctuations might affect the accuracy of the tests, so I launched two versions of the server simultaneously for parallel testing. and the results are as follows.

**memtier_benchmark --data-size 521 --ratio 1:1 --key-pattern R:R --key-minimum=1 --key-maximum 3000000 --test-time 60 -c 500 -t 20 --hide-histogram -x 3**

**before**
```
20        Threads
500       Connections per thread
60        Seconds


BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets        89867.47          ---          ---        56.09141        35.32700       274.43100       350.20700     49902.83 
Gets        89783.04         0.00     89783.04        56.27424        36.09500       274.43100       350.20700      3475.39 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     179650.51         0.00     89783.04        56.18278        35.58300       274.43100       350.20700     53378.23 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets        59275.96          ---          ---        84.31054        39.16700       839.67900      1818.62300     32915.48 
Gets        59193.84         0.00     59193.84        84.22673        39.16700       835.58300      1810.43100      2291.49 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     118469.80         0.00     59193.84        84.26867        39.16700       835.58300      1810.43100     35206.97 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets        79158.44          ---          ---        63.62805        37.37500       376.83100      1294.33500     43956.14 
Gets        79074.76         0.00     79074.76        63.71650        37.88700       376.83100      1294.33500      3060.97 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     158233.20         0.00     79074.76        63.67225        37.63100       376.83100      1294.33500     47017.11
```

**after**
```
20        Threads
500       Connections per thread
60        Seconds


BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       112538.02          ---          ---        44.46570        22.39900       211.96700       303.10300     62490.70 
Gets       112454.14         0.00    112454.14        44.43503        22.27100       211.96700       303.10300      4353.97 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     224992.17         0.00    112454.14        44.45037        22.27100       211.96700       303.10300     66844.67 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets        81328.59          ---          ---        61.54919        24.83100       626.68700      1826.81500     45161.39 
Gets        81243.91         0.00     81243.91        61.62576        24.95900       630.78300      1826.81500      3144.96 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     162572.49         0.00     81243.91        61.58746        24.83100       630.78300      1826.81500     48306.34 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       101919.37          ---          ---        49.09075        23.03900       278.52700      1261.56700     56594.63 
Gets       101836.05         0.00    101836.05        49.07779        22.91100       276.47900      1277.95100      3942.64 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     203755.43         0.00    101836.05        49.08427        22.91100       276.47900      1269.75900     60537.27
```

**memtier_benchmark --data-size 521 --ratio 1:1 --key-pattern R:R --key-minimum=1 --key-maximum 3000000 --test-time 60 -c 50 -t 10 --hide-histogram -x 3**

**before**
```
10        Threads
50        Connections per thread
60        Seconds


BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       178806.91          ---          ---         1.39813         0.83100        30.46300        38.39900     99292.45 
Gets       178803.09       676.94    178126.15         1.39469         0.83100        30.33500        38.39900      7266.89 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     357610.00       676.94    178126.15         1.39641         0.83100        30.46300        38.39900    106559.34 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       173091.03          ---          ---         1.44238         0.83900        30.20700        38.91100     96118.26 
Gets       173087.06      1307.72    171779.34         1.44277         0.83900        30.20700        38.91100      7368.50 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     346178.09      1307.72    171779.34         1.44257         0.83900        30.20700        38.91100    103486.76 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       176391.37          ---          ---         1.41623         0.83100        30.59100        38.91100     97951.03 
Gets       176387.32      1109.75    175277.57         1.41493         0.83100        30.59100        38.91100      7394.89 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     352778.69      1109.75    175277.57         1.41558         0.83100        30.59100        38.91100    105345.92
```

**after**
```
10        Threads
50        Connections per thread
60        Seconds


BEST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       218795.16          ---          ---         1.14316         0.66300        26.23900        38.14300    121498.13 
Gets       218791.07      1867.89    216923.19         1.13932         0.66300        26.11100        38.14300      9423.34 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     437586.23      1867.89    216923.19         1.14124         0.66300        26.23900        38.14300    130921.47 


WORST RUN RESULTS
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       188671.06          ---          ---         1.32398         0.79100        29.82300        37.88700    104770.01 
Gets       188666.88       753.95    187912.93         1.32277         0.79100        29.82300        37.88700      7687.79 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     377337.94       753.95    187912.93         1.32337         0.79100        29.82300        37.88700    112457.79 


AGGREGATED AVERAGE RESULTS (3 runs)
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets       205045.98          ---          ---         1.21886         0.71900        28.54300        38.39900    113863.15 
Gets       205041.80      1436.04    203605.76         1.21660         0.71900        28.41500        38.39900      8670.39 
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals     410087.78      1436.04    203605.76         1.21773         0.71900        28.41500        38.39900    122533.54
```